refactor: 플레이어의 이름을 클래스로 포장
