+++ src/main/java/domain/Name.java
@@ -1,0 +1,42 @@
+package domain;
+
+import static message.ErrorMessage.INVALID_PLAYER_NAME_LANGUAGE_EXCEPTION;
+import static message.ErrorMessage.INVALID_PLAYER_NAME_SIZE_EXCEPTION;
+import static message.ErrorMessage.NO_PLAYER_NAME_EXCEPTION;
+
+public class Name {
+    private final String name;
+
+    public Name(String name) {
+        validate(name);
+        this.name = name;
+    }
+
+    private void validate(String name) {
+        validateNameBlank(name);
+        validateNameLanguage(name);
+        validateNameSize(name);
+    }
+
+    private void validateNameBlank(String name) {
+        if(name.isBlank()){
+            throw new IllegalArgumentException(NO_PLAYER_NAME_EXCEPTION.getMessage());
+        }
+    }
+
+    private void validateNameLanguage(String name) {
+        if(!name.matches("^[A-Za-z]*$")){
+            throw new IllegalArgumentException(INVALID_PLAYER_NAME_LANGUAGE_EXCEPTION.getMessage());
+        }
+    }
+
+    private void validateNameSize(String name) {
+        if(name.length() > 5) {
+            throw new IllegalArgumentException(INVALID_PLAYER_NAME_SIZE_EXCEPTION.getMessage());
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+}
